Refactor rte_events functions to static methods
